JBM-281 override getDatasets method
The existing method just copes with sortField and sortDirection in
parallel. If we need to sort on multiple criteria, this method cannot
help. By implementing the method but using sort, we can specify
different fields as described in EBI Search documentation. It means that
we either use sort or combine sortField and sortDirection.